separate courses schema into separate directory
