+++ functions/routes/courses.js
@@ -1,6 +1,5 @@
 // Courses API route
 
-const Joi = require('@hapi/joi');
 const routes = require('express').Router();
 const admin = require('firebase-admin');
 const db = admin.firestore();
@@ -17,43 +16,11 @@ const {
 const wrap = require('../errors/wrap');
 const { OK } = require('../errors/codes');
 
+const { postSchema, putSchema } = require('../models/courses');
+
 // Collection/doc name in Firestore
 const collection = 'courses';
 const docName = 'course';
-
-/**
- * Schematics for Course data
- */
-const courseString      = Joi.string().min(1);  // general non-empty string
-const courseId          = Joi.number().integer().min(0);
-const courseName        = courseString;
-const courseStartDate   = Joi.string().min(4);
-const courseEndDate     = Joi.string().min(4).allow(null);
-const courseDescription = Joi.array().min(1).items(courseString.required());
-const courseImage       = Joi.object({
-                            url:     Joi.string().uri().required(),
-                            caption: Joi.string().allow("").required(),
-                          });
-const courseImages      = Joi.array().items(courseImage);
-
-// POST /courses schema
-const postSchema = Joi.object({
-  id:          courseId.required(),
-  name:        courseName.required(),
-  description: courseDescription.required(),
-  images:      courseImages.required(),
-  startDate:   courseStartDate.required(),
-  endDate:     courseEndDate.required(),
-});
-
-// PUT /courses/:id schema
-const putSchema = Joi.object({
-  name:        courseName,
-  description: courseDescription,
-  images:      courseImages,
-  startDate:   courseStartDate,
-  endDate:     courseEndDate,
-});
 
 /**
  * Courses endpoints
